requirements.txt
asgiref
certifi
cffi
chardet
cryptography
defusedxml
Django
django-environ
idna
oauthlib
pip
psycopg2
pycparser
python3-openid
pytz
requests==2.25.1
requests-oauthlib
setuptools
django-environ
sqlparse
urllib3
wheel
gunicorn
pillow
sorl-thumbnail==12.6.3
django-autoslug==1.9.8
djangorestframework
django-filter
isort
flake8
